Enable influx bundle to use either FluxQL or InfluxQL language to access an influx database. * add a configuration option to backend/timedata/influx/Config.java (with configuration backward compatibility) * add a configuration option to edge/timedata/influxdb/Config.java (with configuration backward compatibility) * io.openems/shared/influxdb ** splitted into an Interface InfluxConnector ** a Common and a FluxQL implementation ** added a InfluxQL implementation Notes: * provide a getLatestValue implementation also * minor modification to InfluxConnectorFluxQL:queryChannelValuesMap: made hardcoded DB "db/default" configurable, increased data fetch range from "-5m" to "-16m" (helps displaying systems which update only every 15min).
